chore: Moves Tabs to the components folder
