+++ superset-frontend/spec/javascripts/datasource/DatasourceEditor_spec.jsx
@@ -24,7 +24,7 @@ import thunk from 'redux-thunk';
 import { Radio } from 'src/components/Radio';
 
 import Icon from 'src/components/Icon';
-import Tabs from 'src/common/components/Tabs';
+import Tabs from 'src/components/Tabs';
 import DatasourceEditor from 'src/datasource/DatasourceEditor';
 import Field from 'src/CRUD/Field';
 import mockDatasource from 'spec/fixtures/mockDatasource';
