+++ superset-frontend/src/common/components/common.stories.tsx
@@ -23,7 +23,7 @@ import { CronPicker, CronError } from 'src/components/CronPicker';
 import Modal from 'src/components/Modal';
 import InfoTooltip from 'src/components/InfoTooltip';
 import { Dropdown } from 'src/components/Dropdown';
-import Tabs, { EditableTabs } from './Tabs';
+import Tabs, { EditableTabs } from 'src/components/Tabs';
 import { Menu, Input, Divider } from '.';
 
 export default {
